HDDS-2728. Remove methods of internal representation from DatanodeAdminMontor interface
