--- hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/node/DatanodeAdminMonitorInterface.java
@@ -1,43 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.hadoop.hdds.scm.node;
-
-import org.apache.hadoop.hdds.conf.OzoneConfiguration;
-import org.apache.hadoop.hdds.protocol.DatanodeDetails;
-import org.apache.hadoop.hdds.scm.container.ReplicationManager;
-import org.apache.hadoop.hdds.scm.pipeline.PipelineManager;
-import org.apache.hadoop.hdds.server.events.EventPublisher;
-
-/**
- * Interface used by the DatanodeAdminMonitor, which can be used to
- * decommission or recommission nodes and take them in and out of maintenance.
- */
-public interface DatanodeAdminMonitorInterface extends Runnable {
-
-  void setConf(OzoneConfiguration conf);
-  void setEventQueue(EventPublisher scm);
-  void setNodeManager(NodeManager nm);
-  void setPipelineManager(PipelineManager pm);
-  void setReplicationManager(ReplicationManager rm);
-  void startMonitoring(DatanodeDetails dn, int endInHours);
-  void stopMonitoring(DatanodeDetails dn);
-
-  int getPendingCount();
-  int getCancelledCount();
-  int getTrackedNodeCount();
-}
